+++ apps/web/nerdwork-plus/app/(protected)/(creator)/_components/comics/ChaptersEmptyState.tsx
@@ -1,7 +1,7 @@
 import { Book, Plus } from "lucide-react";
 import Link from "next/link";
 import React from "react";
-import { Button } from "../../../../components/ui/button";
+import { Button } from "@/components/ui/button";
 
 const ChaptersEmptyState = ({ comicId }: { comicId: number }) => {
   return (
