+++ apps/web/nerdwork-plus/app/(protected)/(creator)/_components/comics/ComicsHeader.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Button } from "../../../../components/ui/button";
+import { Button } from "@/components/ui/button";
 import { Plus } from "lucide-react";
 import Link from "next/link";
