+++ apps/web/nerdwork-plus/app/(protected)/(creator)/_components/nfts/NFTsEmptyState.tsx
@@ -2,7 +2,7 @@
 import { Book, Plus } from "lucide-react";
 // import Link from "next/link";
 import React from "react";
-import { Button } from "../../../../components/ui/button";
+import { Button } from "@/components/ui/button";
 import { toast } from "sonner";
 
 const NFTsEmptyState = () => {
