+++ apps/web/nerdwork-plus/app/(protected)/(creator)/creator/comics/page.tsx
@@ -1,9 +1,9 @@
-import Comics from "@/app/(creator)/_components/comics/Comics";
-import ComicsHeader from "@/app/(creator)/_components/comics/ComicsHeader";
-import MyComicsEmptyState from "@/app/(creator)/_components/comics/MyComicsEmptyState";
 import { comicData } from "@/components/data";
 import { Metadata } from "next";
 import React from "react";
+import ComicsHeader from "../../_components/comics/ComicsHeader";
+import MyComicsEmptyState from "../../_components/comics/MyComicsEmptyState";
+import Comics from "../../_components/comics/Comics";
 
 export const metadata: Metadata = {
   title: "Creator Dashboard",
